Make in_length member unsigned, so negative values are detected at compile time
The buffer length member in libusb is of type int, but must be
non-negative. This implements a simple compile-time check to fulfill that constraint.

+++ src/controllers/bulk/bulkcontroller.cpp
@@ -301,7 +301,7 @@ int BulkController::open(const QString& resourcePath) {
         return -1;
     }
     m_inEndpointAddr = pDevice->endpoints.in_epaddr;
-    m_inLength = pDevice->endpoints.in_length;
+    m_inLength = static_cast<int>(pDevice->endpoints.in_length);
     m_outEndpointAddr = pDevice->endpoints.out_epaddr;
     m_interfaceNumber = pDevice->endpoints.interface_number;
